requirements.txt
asn1crypto==1.3.0
backports.csv==1.0.7
bcrypt==3.1.7
bleach==3.1.5
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.7
decorator==4.4.0
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==2.0.5
django-heroku==0.3.1
django-import-export==1.2.0
django-storages==1.9.1
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
docutils==0.16
dotmap==1.3.8
et-xmlfile==1.0.1
Fabric3==1.14.post1
gunicorn==19.7.1
idna==2.8
importlib-metadata==1.3.0
jdcal==1.4.1
keyring==21.0.0
more-itertools==8.0.2
odfpy==1.4.0
openpyxl==3.0.3
packaging==20.4
paramiko==2.7.1
pkginfo==1.5.0.1
prompt-toolkit==2.0.9
psycopg2==2.8.4
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
PyNaCl==1.3.0
pyparsing==2.4.7
python-decouple==3.3
pytz==2019.3
pywin32-ctypes==0.2.0
PyYAML==5.1.2
readme-renderer==24.0
requests==2.23.0
requests-toolbelt==0.9.1
sentry-sdk==0.14.2
six==1.12.0
tablib==0.13.0
tqdm==4.41.1
twine==3.1.1
urllib3==1.25.8
wcwidth==0.2.4
webencodings==0.5.1
whitenoise==4.1.4
wincertstore==0.2
xlrd==1.2.0
xlwt==1.3.0
zipp==0.6.0
